test: move bundle deps test to separate file
